finished backend, finished frontend template. Start Jquery/ajax and add user picture
