+++ Server/Routes/index.js
@@ -1,6 +1,7 @@
 const routes = require('express').Router()
 const userRoutes = require('./user')
 const todoRoutes = require('./todo')
+const projectRoutes = require('./project')
 
 
 routes.get('/', (req, res) => {
@@ -10,5 +11,6 @@ routes.get('/', (req, res) => {
 
 routes.use('/', userRoutes)
 routes.use('/', todoRoutes)
+routes.use('/', projectRoutes)
 
 module.exports = routes
